Lag felles funksjonalitet for å utlede første steg på en behandling
